Added logging and made the code modular and pretty.
Added a function which will show progress bar while moving the existing
content with the installtion of the new wordpress

Also added a logging functionality which will log every commad made by
the user

Added Trap flags which is mainly focused to trap the exit operation
while upgrading or with the installation of the wordpress. This will
clean up the files.

Added code to check the nginx status and also restart on prompt if the
nginx configuration is alright.