Added a Database file and fixed bug with the Nginx module
The database module will be used to alter and get database. This will later be used to ensure that installation command will do everything like configuring the datase also.

Also the bug with the nginx Module was also resolved.

Next, the installation module will have the feature to create DB.